Split the risk digest formatter from its per-indicator lines
Pure move: IndicatorDetailFormatter owns the published-indicator and
outcome line formatting, while RiskIndicatorDigestFormatter keeps the
digest assembly and delegates per-item rendering; the digest wire text
stays byte-identical and the assembly tests run against the real detail
formatter rather than a mock.

God classes 11 -> 10, PMD violations 147 -> 144, top-15 lines
7050 -> 6876.
