+++ metrics/scorecard.json
@@ -1,5 +1,5 @@
 {
-  "branchCoverage": 85.96,
+  "branchCoverage": 85.97,
   "classCouplingViolations": 35,
   "clockViolations": 0,
   "cognitiveComplexityViolations": 20,
@@ -7,9 +7,9 @@
   "deepNestingViolations": 0,
   "disabledTests": 5,
   "disconnectedClasses": 454,
-  "godClasses": 11,
-  "lineCoverage": 82.95,
-  "longClasses": 4,
+  "godClasses": 10,
+  "lineCoverage": 82.96,
+  "longClasses": 3,
   "meanInstability": 0.473,
   "moduleCycles": 0,
   "modulithViolations": 0,
@@ -18,19 +18,19 @@
   "mutationsSliced": 2339,
   "mutationsTotal": 12781,
   "pitestSlicedModules": 10,
-  "pmdViolations": 147,
+  "pmdViolations": 144,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
     "CouplingBetweenObjects": 35,
-    "CyclomaticComplexity": 76,
-    "GodClass": 11,
-    "TooManyMethods": 5
+    "CyclomaticComplexity": 75,
+    "GodClass": 10,
+    "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 113.2,
+  "slowest10TestClassesSeconds": 111.0,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 181,
+  "suiteWallTimeSeconds": 171,
   "testClassesOver10Seconds": 5,
-  "top15ClassLines": 7050,
+  "top15ClassLines": 6876,
   "totalPackages": 227,
   "unmarkedPackages": 0
 }
