+++ src/main/java/ee/tuleva/onboarding/investment/risk/RiskIndicatorDigestFormatter.java
@@ -1,22 +1,13 @@
 package ee.tuleva.onboarding.investment.risk;
 
-import static ee.tuleva.onboarding.investment.risk.RiskIndicatorType.SRI;
-
 import ee.tuleva.onboarding.comparisons.fundvalue.FundValueQueries;
-import ee.tuleva.onboarding.investment.risk.RiskIndicatorService.RiskIndicatorOutcome;
 import ee.tuleva.onboarding.investment.risk.RiskIndicatorService.RiskIndicatorRun;
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.time.Clock;
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.EnumMap;
 import java.util.List;
 import java.util.Locale;
-import java.util.Objects;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.Nullable;
 import org.springframework.stereotype.Component;
@@ -25,14 +16,11 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 class RiskIndicatorDigestFormatter {
 
-  private static final int PERSISTENCE_WINDOW_MONTHS = 4;
-  private static final int FULL_SRRI_WINDOW_OBSERVATIONS = 260;
   private static final Locale ESTONIAN = Locale.of("et", "EE");
 
-  private final DisclosedRiskIndicatorRepository disclosureRepository;
   private final RiskIndicatorProperties properties;
   private final FundValueQueries fundValueQueries;
-  private final Clock clock;
+  private final IndicatorDetailFormatter detailFormatter;
 
   String digest(RiskIndicatorRun run) {
     var proxyReviews = proxyReviewLines(run);
@@ -44,7 +32,7 @@ class RiskIndicatorDigestFormatter {
     message.append("```\n").append(table(run)).append("```\n");
 
     for (var outcome : run.outcomes()) {
-      message.append("\n").append(detailBlock(outcome)).append("\n");
+      message.append("\n").append(detailFormatter.detailBlock(outcome)).append("\n");
     }
 
     proxyReviews.forEach(line -> message.append("\n").append(line).append("\n"));
@@ -60,13 +48,12 @@ class RiskIndicatorDigestFormatter {
 
   private String headline(RiskIndicatorRun run, List<String> proxyReviews) {
     var counts = new EnumMap<Severity, Integer>(Severity.class);
-    run.outcomes()
-        .forEach(
-            outcome ->
-                counts.merge(
-                    severity(outcome.indicator(), disclosedClass(outcome.indicator())),
-                    1,
-                    Integer::sum));
+    for (var outcome : run.outcomes()) {
+      var severity =
+          detailFormatter.severity(
+              outcome.indicator(), detailFormatter.disclosedClass(outcome.indicator()));
+      counts.merge(severity, 1, Integer::sum);
+    }
 
     var parts = new ArrayList<String>();
     if (counts.getOrDefault(Severity.RED, 0) > 0) {
@@ -96,10 +83,16 @@ class RiskIndicatorDigestFormatter {
   }
 
   private Severity worstSeverityAcrossFunds(RiskIndicatorRun run) {
-    return run.outcomes().stream()
-        .map(outcome -> severity(outcome.indicator(), disclosedClass(outcome.indicator())))
-        .max(Comparator.naturalOrder())
-        .orElse(Severity.GREEN);
+    var worst = Severity.GREEN;
+    for (var outcome : run.outcomes()) {
+      var severity =
+          detailFormatter.severity(
+              outcome.indicator(), detailFormatter.disclosedClass(outcome.indicator()));
+      if (severity.compareTo(worst) > 0) {
+        worst = severity;
+      }
+    }
+    return worst;
   }
 
   private boolean hasLinesTiedToNoFund(RiskIndicatorRun run, List<String> proxyReviews) {
@@ -130,269 +123,20 @@ class RiskIndicatorDigestFormatter {
             "Staatus"));
     for (var outcome : run.outcomes()) {
       var indicator = outcome.indicator();
-      var disclosed = disclosedClass(indicator);
+      var row = detailFormatter.tableRow(indicator);
       table.append(
           String.format(
               format,
               indicator.fund(),
               indicator.indicatorType(),
               text(indicator.publishedClass()),
-              disclosed == null ? "?" : String.valueOf(disclosed.getDisclosedClass()),
-              publishedSince(indicator),
-              duration(indicator),
+              row.disclosedClassColumn(),
+              row.publishedSince(),
+              row.duration(),
               text(indicator.previousPublishedClass()),
-              statusLabel(indicator, disclosed)));
+              row.statusLabel()));
     }
     return table.toString();
-  }
-
-  private String detailBlock(RiskIndicatorOutcome outcome) {
-    var indicator = outcome.indicator();
-    var disclosed = disclosedClass(indicator);
-    var block = new ArrayList<String>();
-
-    if (!indicator.hasClass()) {
-      block.add("⚠️ %s %s — andmeid napib".formatted(indicator.fund(), indicator.indicatorType()));
-      block.add(
-          "Aknas on %d vaatlust, klassi ei avaldata. Volatiilsus %s."
-              .formatted(indicator.latestObservationCount(), volatility(indicator)));
-      block.add(
-          "👉 Tegevus: kontrolli, kas NAV-seeria on täielik. Kui fondi enda ajalugu ongi nõutud"
-              + " perioodist lühem, lisa investment.risk.sources alla võrdlusindeksi segment —"
-              + " klassi avaldamata jätmine ei ole lubatud variant.");
-      return String.join("\n", withDiagnostics(block, outcome));
-    }
-
-    if (isMismatched(disclosed, indicator)) {
-      block.add(mismatchLine(indicator, disclosed));
-      block.add(
-          "Muutus jõustus %s. Viimane dokument: '%s' (klass %d, alates %s)."
-              .formatted(
-                  indicator.publishedSince(),
-                  Objects.requireNonNull(disclosed).getDocument(),
-                  disclosed.getDisclosedClass(),
-                  disclosed.getDisclosedFrom()));
-      block.add(
-          "👉 Tegevus: dokument vajab uuendamist. Pärast avaldamist lisa rida"
-              + " investment_risk_indicator_disclosure tabelisse.");
-    } else if (disclosed == null) {
-      block.add(
-          "⚠️ %s %s — avaldatud klass teadmata"
-              .formatted(indicator.fund(), indicator.indicatorType()));
-      block.add(
-          "Arvutatud avaldatav klass on %d, aga ühtegi dokumendirida ei ole."
-              .formatted(indicator.publishedClass()));
-      block.add(
-          "👉 Tegevus: lisa kehtiv KID/KIID rida investment_risk_indicator_disclosure"
-              + " tabelisse.");
-    } else if (indicator.status() == RiskIndicatorStatus.CHANGE_PENDING) {
-      block.add("⚠️ %s %s — muutus ootel".formatted(indicator.fund(), indicator.indicatorType()));
-      block.add(
-          "Toores klass %s, avaldatav klass %d. %s %s."
-              .formatted(
-                  indicator.rawLatestClass(),
-                  indicator.publishedClass(),
-                  volatilityLabel(indicator),
-                  volatility(indicator)));
-      block.add(pendingProgress(indicator));
-      block.add("👉 Tegevus praegu pole — jälgi.");
-    } else if (indicator.status() == RiskIndicatorStatus.CHANGE_CONFIRMED) {
-      block.add(
-          "⚠️ %s %s — muutus äsja kinnitatud"
-              .formatted(indicator.fund(), indicator.indicatorType()));
-      block.add(
-          "Avaldatav klass %d alates %s (eelmine %s). %s"
-              .formatted(
-                  indicator.publishedClass(),
-                  indicator.publishedSince(),
-                  text(indicator.previousPublishedClass()),
-                  rangeLine(indicator)));
-      block.add("👉 Tegevus: kontrolli, kas dokument on juba uuendatud.");
-    } else {
-      block.add(
-          "✅ %s %s — stabiilne, dokument ajakohane"
-              .formatted(indicator.fund(), indicator.indicatorType()));
-      block.add(rangeLine(indicator));
-      block.add(majorityLine(indicator));
-    }
-
-    return String.join("\n", withDiagnostics(block, outcome));
-  }
-
-  private List<String> withDiagnostics(List<String> block, RiskIndicatorOutcome outcome) {
-    dataQualityLine(outcome.indicator()).ifPresent(block::add);
-    truncatedHistoryLine(outcome.indicator()).ifPresent(block::add);
-    driftLine(outcome).ifPresent(block::add);
-    redefinitionLine(outcome).ifPresent(block::add);
-    skippedLine(outcome).ifPresent(block::add);
-    return block;
-  }
-
-  String mismatchLine(
-      PublishedRiskIndicator indicator, @Nullable DisclosedRiskIndicator disclosed) {
-    return "🔴 %s %s — dokumendis on klass %s, arvutatud avaldatav klass on %s"
-        .formatted(
-            indicator.fund(),
-            indicator.indicatorType(),
-            disclosed == null ? "?" : disclosed.getDisclosedClass(),
-            indicator.publishedClass());
-  }
-
-  private String rangeLine(PublishedRiskIndicator indicator) {
-    var publishedClass = Objects.requireNonNull(indicator.publishedClass());
-    var range = RiskClassBucket.range(indicator.indicatorType(), publishedClass);
-    var volatility = indicator.latestVolatility();
-    if (volatility == null) {
-      return "Klassi %d vahemik %s.".formatted(publishedClass, rangeText(indicator, range));
-    }
-    var distance =
-        RiskClassBucket.distanceToNearestBound(
-            indicator.indicatorType(), publishedClass, volatility);
-    return "%s %s (klassi %d vahemik %s); lähim piir on %s kaugusel."
-        .formatted(
-            volatilityLabel(indicator),
-            volatility(indicator),
-            publishedClass,
-            rangeText(indicator, range),
-            distance == null ? "—" : number(distance, indicator));
-  }
-
-  private String rangeText(PublishedRiskIndicator indicator, RiskClassBucket.ClassRange range) {
-    var lower = range.lowerInclusive();
-    var upper = range.upperExclusive();
-    return "%s–%s"
-        .formatted(
-            lower == null ? "0" : number(lower, indicator),
-            upper == null ? "∞" : number(upper, indicator));
-  }
-
-  private String majorityLine(PublishedRiskIndicator indicator) {
-    if (indicator.indicatorType() == SRI) {
-      var needed = indicator.windowReferencePoints() / 2 + 1;
-      return "PRIIPs 4-kuu enamus: %d/%d referentspunkti klassile %s; pöördeks on vaja %d"
-              .formatted(
-                  indicator.matchingReferencePoints(),
-                  indicator.windowReferencePoints(),
-                  text(indicator.rawLatestClass()),
-                  needed)
-          + " vastupidist referentspunkti.";
-    }
-    return "CESR 4-kuu aken: %d/%d nädalat klassile %s; migratsiooniks peab volatiilsus olema kõik"
-            .formatted(
-                indicator.matchingReferencePoints(),
-                indicator.windowReferencePoints(),
-                text(indicator.rawLatestClass()))
-        + " neli kuud väljaspool avaldatavat klassi.";
-  }
-
-  private String pendingProgress(PublishedRiskIndicator indicator) {
-    var since = indicator.rawClassSince();
-    if (indicator.indicatorType() == SRI) {
-      var needed = indicator.windowReferencePoints() / 2 + 1;
-      var missing = Math.max(0, needed - indicator.matchingReferencePoints());
-      return "Klass %s on hoidnud %d kauplemispäeva alates %s. PRIIPs enamuseni puudu %d punkti."
-          .formatted(
-              text(indicator.rawLatestClass()),
-              indicator.rawStreakReferencePoints(),
-              text(since),
-              missing);
-    }
-    if (since == null) {
-      return "Klass %s on hoidnud %d nädalat."
-          .formatted(text(indicator.rawLatestClass()), indicator.rawStreakReferencePoints());
-    }
-    var threshold = since.plusMonths(PERSISTENCE_WINDOW_MONTHS);
-    var weeksLeft =
-        Math.max(0, ChronoUnit.WEEKS.between(indicator.evaluationDate(), threshold) + 1);
-    return "Klass %s on püsinud %d nädalat alates %s. CESR 4-kuu künniseni puudu %d nädalat,"
-            .formatted(
-                text(indicator.rawLatestClass()),
-                indicator.rawStreakReferencePoints(),
-                since,
-                weeksLeft)
-        + " eeldatav kinnitus %s; aknas on veel %d referentspunkti muus klassis."
-            .formatted(threshold, referencePointsNotYetOnTheRawClass(indicator));
-  }
-
-  private int referencePointsNotYetOnTheRawClass(PublishedRiskIndicator indicator) {
-    return indicator.windowReferencePoints() - indicator.matchingReferencePoints();
-  }
-
-  private Optional<String> truncatedHistoryLine(PublishedRiskIndicator indicator) {
-    if (!indicator.publishedSinceIsTruncated()) {
-      return Optional.empty();
-    }
-    return Optional.of(
-        "⚠️ %s %s: jooks algab salvestatud seeria esimesest punktist (%s), seega \"kehtib alates\""
-                .formatted(indicator.fund(), indicator.indicatorType(), indicator.publishedSince())
-            + " on alampiir, mitte tegelik algus — jooksuta RiskIndicatorBackfillJob.");
-  }
-
-  private Optional<String> driftLine(RiskIndicatorOutcome outcome) {
-    var drifted = outcome.driftedDates();
-    if (drifted.isEmpty()) {
-      return Optional.empty();
-    }
-    return Optional.of(
-        "⚠️ %s %s: %d varasemat referentspunkti arvutati ümber (vanim %s, viimane %s) —"
-                .formatted(
-                    outcome.indicator().fund(),
-                    outcome.indicator().indicatorType(),
-                    drifted.size(),
-                    drifted.getFirst(),
-                    drifted.getLast())
-            + " allikaandmed muutusid tagantjärele.");
-  }
-
-  private Optional<String> redefinitionLine(RiskIndicatorOutcome outcome) {
-    var redefinitions = outcome.redefinitions();
-    if (redefinitions.isEmpty()) {
-      return Optional.empty();
-    }
-    var first = redefinitions.getFirst();
-    return Optional.of(
-        "ℹ️ %s %s: %d varasemat referentspunkti arvutati ümber (vanim %s, viimane %s) —"
-                .formatted(
-                    outcome.indicator().fund(),
-                    outcome.indicator().indicatorType(),
-                    redefinitions.size(),
-                    first.date(),
-                    redefinitions.getLast().date())
-            + " hoidmisperioodi kauplemispäevi %s → %s. Alusandmed ei muutunud."
-                .formatted(
-                    text(first.previousHoldingPeriodTradingDays()),
-                    first.holdingPeriodTradingDays()));
-  }
-
-  private Optional<String> skippedLine(RiskIndicatorOutcome outcome) {
-    var skipped = outcome.skippedDates();
-    if (skipped.isEmpty()) {
-      return Optional.empty();
-    }
-    return Optional.of(
-        "⚠️ %s %s: %d referentspunkti jäi arvutamata (vanim %s, viimane %s) — VEV ei tulnud"
-                .formatted(
-                    outcome.indicator().fund(),
-                    outcome.indicator().indicatorType(),
-                    skipped.size(),
-                    skipped.getFirst(),
-                    skipped.getLast())
-            + " lõplik arv, seeriasse jäi auk. 👉 Tegevus: kontrolli nende kuupäevade alushindu.");
-  }
-
-  private Optional<String> dataQualityLine(PublishedRiskIndicator indicator) {
-    if (indicator.indicatorType() == SRI
-        || indicator.latestObservationCount() >= FULL_SRRI_WINDOW_OBSERVATIONS
-        || indicator.latestObservationCount() == 0) {
-      return Optional.empty();
-    }
-    return Optional.of(
-        "⚠️ %s %s: 5a aknas %d/%d nädalatootlust — puuduv NAV tuleb index_values tabelis parandada."
-            .formatted(
-                indicator.fund(),
-                indicator.indicatorType(),
-                indicator.latestObservationCount(),
-                FULL_SRRI_WINDOW_OBSERVATIONS));
   }
 
   private List<String> proxyReviewLines(RiskIndicatorRun run) {
@@ -432,111 +176,16 @@ class RiskIndicatorDigestFormatter {
   }
 
   private String footer(RiskIndicatorRun run) {
-    var sources =
-        run.outcomes().stream()
-            .map(
-                outcome ->
-                    "%s: %s"
-                        .formatted(
-                            outcome.indicator().fund(),
-                            String.join(
-                                "+",
-                                properties.sourcesFor(outcome.indicator().fund()).stream()
-                                    .map(RiskIndicatorProperties.Source::key)
-                                    .toList())))
-            .toList();
+    var sources = new ArrayList<String>();
+    for (var outcome : run.outcomes()) {
+      var sourceKeys = new ArrayList<String>();
+      for (var source : properties.sourcesFor(outcome.indicator().fund())) {
+        sourceKeys.add(source.key());
+      }
+      sources.add("%s: %s".formatted(outcome.indicator().fund(), String.join("+", sourceKeys)));
+    }
     return "Allikad: %s. SRI = MRM, eeldusel CRM = %d.%n"
         .formatted(String.join("; ", sources), SriCalculator.ASSUMED_CREDIT_RISK_MEASURE);
-  }
-
-  @Nullable DisclosedRiskIndicator disclosedClass(PublishedRiskIndicator indicator) {
-    return disclosureRepository
-        .findFirstByIndicatorTypeAndFundAndDisclosedFromLessThanEqualOrderByDisclosedFromDesc(
-            indicator.indicatorType(), indicator.fund(), LocalDate.now(clock))
-        .orElse(null);
-  }
-
-  boolean isMismatched(
-      @Nullable DisclosedRiskIndicator disclosed, PublishedRiskIndicator indicator) {
-    return disclosed != null
-        && indicator.publishedClass() != null
-        && !Objects.equals(disclosed.getDisclosedClass(), indicator.publishedClass());
-  }
-
-  Severity severity(PublishedRiskIndicator indicator, @Nullable DisclosedRiskIndicator disclosed) {
-    if (isMismatched(disclosed, indicator)) {
-      return Severity.RED;
-    }
-    if (!indicator.hasClass() || disclosed == null) {
-      return Severity.YELLOW;
-    }
-    return indicator.status() == RiskIndicatorStatus.STABLE ? Severity.GREEN : Severity.YELLOW;
-  }
-
-  private String statusLabel(
-      PublishedRiskIndicator indicator, @Nullable DisclosedRiskIndicator disclosed) {
-    var icon = severity(indicator, disclosed).icon();
-    if (!indicator.hasClass()) {
-      return icon + " andmeid napib";
-    }
-    if (isMismatched(disclosed, indicator)) {
-      return icon + " DOKUMENT VANANENUD";
-    }
-    if (disclosed == null) {
-      return icon + " avaldatud teadmata";
-    }
-    return icon
-        + switch (indicator.status()) {
-          case STABLE -> " stabiilne";
-          case CHANGE_PENDING -> " muutus ootel";
-          case CHANGE_CONFIRMED -> " muutus kinnitatud";
-        };
-  }
-
-  private String publishedSince(PublishedRiskIndicator indicator) {
-    var since = indicator.publishedSince();
-    if (since == null) {
-      return "—";
-    }
-    return indicator.publishedSinceIsTruncated() ? "≥" + since : since.toString();
-  }
-
-  private String duration(PublishedRiskIndicator indicator) {
-    var since = indicator.publishedSince();
-    if (since == null) {
-      return "—";
-    }
-    var months = ChronoUnit.MONTHS.between(since, indicator.evaluationDate());
-    return months < 12
-        ? "%d kuud".formatted(months)
-        : "%da %dk".formatted(months / 12, months % 12);
-  }
-
-  private String volatilityLabel(PublishedRiskIndicator indicator) {
-    return indicator.indicatorType() == SRI ? "VEV" : "Aastane volatiilsus";
-  }
-
-  private String volatility(PublishedRiskIndicator indicator) {
-    var volatility = indicator.latestVolatility();
-    return volatility == null ? "—" : number(volatility, indicator);
-  }
-
-  private String number(BigDecimal value, PublishedRiskIndicator indicator) {
-    return indicator.indicatorType() == SRI
-        ? vevInDecimals(value)
-        : annualisedVolatilityInPerCent(value);
-  }
-
-  private String vevInDecimals(BigDecimal value) {
-    return String.format(ESTONIAN, "%.4f", value);
-  }
-
-  private String annualisedVolatilityInPerCent(BigDecimal value) {
-    return String.format(
-            ESTONIAN,
-            "%.2f",
-            value.multiply(BigDecimal.valueOf(100)).setScale(2, RoundingMode.HALF_UP))
-        + "%";
   }
 
   private String text(@Nullable Object value) {
